# fix ClassTag#cache leak