I was diagnosing some classloader leaks and the profiler pointed to `ClassTag#cache`

JDK issues about the problem:
* https://bugs.openjdk.java.net/browse/JDK-8136353

Also this is not a problem in 2.13.x because the cache does not exist there. See gitter question https://gitter.im/scala/contributors?at=5f941163a0a3072b43a099b7