@jtjeferreira 2020-10-24
Would be interesting to write a test similar to the one in the JDK bug:

```
public static void main(String[] args) throws Throwable {
        for (long i = 0; i<10000000; i++) {
            File dir = new File("t/t.jar");
            URLClassLoader classLoader = new URLClassLoader(new URL[]{dir.toURI().toURL()});
            ClassValue cv = (ClassValue) classLoader.loadClass("MyClassValue").newInstance();
            Object value = cv.get(Integer.TYPE);
            assert value !=null;
            assert value.getClass().getClassLoader() == classLoader;
            classLoader.close();
        }

    } 
```

but I would need some help since this is my first contribution :joy: