@lrytz 2020-10-26
@retronym do you know why the cache doesn't exist in 2.13.x?

It was originally added for 2.12.x in https://github.com/scala/scala/pull/7879 and then merged forward in https://github.com/scala/scala/pull/8242, where you marked it as "PICK", but it doesn't show up in the diff.