@retronym 2020-10-27
Thanks for bringing up this issue.

An alternative fix might be to limit our use of the cache to when `val cl = runtimeClass.getClassLoader; cl == null || cl == classOf[scala.reflect.runtime.universe.type].getClassLoader`. Classes loaded from child classloaders would not have entries for these caches installed in their `ClassValue` maps.

We use `ClassValue` in a few other places; they will also need an update.

I'll try out some variations of the fix and also try to replicate the leak in a test.