@retronym 2020-10-27
Related fix for `TypeTag` reported in https://github.com/scala/bug/issues/11767 and fixed in https://github.com/scala/bug/issues/11767 in a manner similar to your proposal.