@jtjeferreira 2020-10-27
@retronym looking forward for your variations of the fix, but "do you know why the cache doesn't exist in 2.13.x?" as @lrytz asked? Removing the cache in 2.12.x could also be a fix :)