@retronym 2020-11-02
@jtjeferreira Could you describe the classloader heirarchy that showed up the problem? Is `scala-library` being loaded in temporary classloaders and leaking its `ClassValue` in the `Class.classValueMap` of classes from parent loaders? Or is a single load of `scala-library` that is used to form `ClassTag`s over a user class in temporary/child classloader preventing unloading of that user class? I can't get any test to show a leak, but if I know which one I'm going for should be able to repro.