@retronym 2020-11-02
I've pushed a commit to deal with the race condition of the weak reference immediately getting cleared. I do the same as we do in `TypeTag`'s use of `ClassValue`/`WeakReference` and just return an uncached result after removing a stale cache entry. 

Baseline:

```
[info] Benchmark                                             Mode  Cnt   Score    Error   Units
[info] ClassTagBenchmark.lookupClassTag                      avgt   20  40.625 ±  2.595   ns/op
[info] ClassTagBenchmark.lookupClassTag:·gc.alloc.rate       avgt   20   0.001 ±  0.002  MB/sec
[info] ClassTagBenchmark.lookupClassTag:·gc.alloc.rate.norm  avgt   20  ≈ 10⁻⁴             B/op
[info] ClassTagBenchmark.lookupClassTag:·gc.count            avgt   20     ≈ 0           counts
```

This PR:

```
[info] Benchmark                                             Mode  Cnt   Score    Error   Units
[info] ClassTagBenchmark.lookupClassTag                      avgt   20  47.246 ±  4.295   ns/op
[info] ClassTagBenchmark.lookupClassTag:·gc.alloc.rate       avgt   20   0.001 ±  0.002  MB/sec
[info] ClassTagBenchmark.lookupClassTag:·gc.alloc.rate.norm  avgt   20  ≈ 10⁻⁴             B/op
[info] ClassTagBenchmark.lookupClassTag:·gc.count            avgt   20     ≈ 0           counts
```

With the cache disabled altogether:

```
[info] Benchmark                                                          Mode  Cnt     Score     Error   Units
[info] ClassTagBenchmark.lookupClassTag                                   avgt   20    48.622 ±   2.184   ns/op
[info] ClassTagBenchmark.lookupClassTag:·gc.alloc.rate                    avgt   20  2933.149 ± 131.283  MB/sec
[info] ClassTagBenchmark.lookupClassTag:·gc.alloc.rate.norm               avgt   20   224.000 ±   0.001    B/op
[info] ClassTagBenchmark.lookupClassTag:·gc.churn.PS_Eden_Space           avgt   20  2930.875 ± 352.368  MB/sec
[info] ClassTagBenchmark.lookupClassTag:·gc.churn.PS_Eden_Space.norm      avgt   20   223.683 ±  24.055    B/op
[info] ClassTagBenchmark.lookupClassTag:·gc.churn.PS_Survivor_Space       avgt   20     0.097 ±   0.038  MB/sec
[info] ClassTagBenchmark.lookupClassTag:·gc.churn.PS_Survivor_Space.norm  avgt   20     0.007 ±   0.003    B/op
[info] ClassTagBenchmark.lookupClassTag:·gc.count                         avgt   20   119.000            counts
[info] ClassTagBenchmark.lookupClassTag:·gc.time                          avgt   20   363.000                ms
```

In this microbenchmark, the weak reference version has similar execution time to the uncached version. But it does reduce allocation rates.