@jtjeferreira 2020-11-02
> @jtjeferreira Could you describe the classloader heirarchy that showed up the problem? Is `scala-library` being loaded in temporary classloaders and leaking its `ClassValue` in the `Class.classValueMap` of classes from parent loaders? Or is a single load of `scala-library` that is used to form `ClassTag`s over a user class in temporary/child classloader preventing unloading of that user class? I can't get any test to show a leak, but if I know which one I'm going for should be able to repro.

This is from a play 2.7.x project when running the application. Each time the application is stopped and started new classloaders are created. Luckily I have this diagram that I used to understand what was going on:

![image](https://user-images.githubusercontent.com/943051/97861548-d9b7c300-1cfb-11eb-999b-521b165b04a6.png)

Starting from the bottom:
* user classes are on the leaf classloader `reloader.currentApplicationClassLoader (play.runsupport.DelegatedResourcesClassLoader)`
* libraries and `scala-library` are on its great parent `applicationLoader (play.runsupport.NamedURLClassLoader)`
* every time the application is stopped, everything below `commonLoader (play.sbt.PlayCommands$$anon$1)` should be GCd, 

The workaround I used in https://github.com/playframework/playframework/pull/10500 is to move loading of `scala-library` to `play.sbt.PlayCommands$$anon$1` which is cached and survives reloads...