@jtjeferreira 2020-11-02
I also found some interesting discussion here http://mail.openjdk.java.net/pipermail/mlvm-dev/2018-March/006847.html