@jtjeferreira 2020-11-02
I was trying to get you a screenshot from the profiler (from running scripted tests of https://github.com/playframework/playframework/pull/10500 without the fix ), but eclipse MAT does not report it, and my yourkit trial just expired (which AFAICT was the one that was showing it clearly)