@jtjeferreira 2020-11-13
> I wasn't able to reproduce with a test case. I've proceeded with my slight refactor of your original idea, together with a system property to disable the cache altogether.

Thanks for the merge! Do you plan to port this to 2.13?