@mkurz 2020-12-10
> Yeah, I will correct the merge error and take this all over to 2.13.x after 2.13.4 is released.

@retronym did you port this fix and #9361 to the 2.13.x branch yet?